mysite/requirement.txt
adium-theme-ubuntu==0.3.4
arrow==0.10.0
bcrypt==3.1.3
cffi==1.10.0
chardet==2.3.0
cryptography==1.7.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.7
django-admin==1.0.2
django-crispy-forms==1.6.1
django-registration==2.2
django-smart-selects==1.5.2
enum34==1.1.6
f.lux-indicator-applet==1.1.10rc0
futures==3.0.5
gunicorn==19.7.1
idna==2.2
ipaddress==1.0.17
keyring==10.3.1
keyrings.alt==2.2
livestreamer==1.12.2
ndg-httpsclient==0.4.2
olefile==0.44
pexpect==4.2.1
Pillow==4.1.1
ply==3.9
psycopg2==2.7.1
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
pygobject==3.22.0
pyOpenSSL==16.2.0
pystache==0.5.4
python-dateutil==2.6.0
python-librtmp==0.3.0
pyxdg==0.25
requests==2.10.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.10.0
static==1.1.1
static3==0.7.0
unity-lens-photos==1.0
urllib3==1.15.1
whitenoise==3.3.0
